backend/requirements.txt
black==22.1.0
Flask==2.0.3
Flask-Cors==3.0.10
pylint==2.12.2
python-dotenv==0.19.2
textract
newspaper3k
spacy
pytest
gunicorn